requirements.txt
pygame==2.3.0
future~=0.18.2
Pillow~=9.0.1